% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indentation.R
\name{grk_reindent}
\alias{grk_reindent}
\alias{grk_reindent_auto_text}
\alias{grk_reindent_auto_file}
\alias{grk_reindent_auto_dir}
\alias{grk_reindent_auto_pkg}
\alias{grk_reindent_tabs_text}
\alias{grk_reindent_tabs_file}
\alias{grk_reindent_tabs_dir}
\alias{grk_reindent_tabs_pkg}
\alias{grk_reindent_spaces_text}
\alias{grk_reindent_spaces_file}
\alias{grk_reindent_spaces_dir}
\alias{grk_reindent_spaces_pkg}
\title{Re-indent R code using tabs or spaces}
\usage{
grk_reindent_auto_text(text, ...)

grk_reindent_auto_file(path, ...)

grk_reindent_auto_dir(path, ...)

grk_reindent_auto_pkg(pkg = ".", ...)

grk_reindent_tabs_text(text, ...)

grk_reindent_tabs_file(path, ...)

grk_reindent_tabs_dir(path, ...)

grk_reindent_tabs_pkg(pkg = ".", ...)

grk_reindent_spaces_text(text, ..., indent_by = 2L)

grk_reindent_spaces_file(path, ..., indent_by = 2L)

grk_reindent_spaces_dir(path, ..., indent_by = 2L)

grk_reindent_spaces_pkg(pkg = ".", ..., indent_by = 2L)
}
\arguments{
\item{text}{A character vector with text to style.}

\item{...}{Arguments passed on to the \code{style} function.}

\item{path}{A character vector with paths to files to style.}

\item{pkg}{Path to a (subdirectory of an) R package.}

\item{indent_by}{The number of spaces by which to indent the code.}
}
\description{
Quickly convert from spaces to tabs or from tabs to spaces in R code
throught a file, directory or package.
}
\section{Functions}{
\itemize{
\item \code{grk_reindent_auto_text}: Re-indent text using tabs or spaces according to the
RStudio project settings or the \code{grkstyle.use_tabs} option.

\item \code{grk_reindent_auto_file}: Re-indent a file using tabs or spaces according to
the RStudio project settings or the \code{grkstyle.use_tabs} option.

\item \code{grk_reindent_auto_dir}: Re-indent a directory using tabs or spaces according
to the RStudio project settings or the \code{grkstyle.use_tabs} option.

\item \code{grk_reindent_auto_pkg}: Re-indent a package using tabs or spaces according
to the RStudio project settings or the \code{grkstyle.use_tabs} option.

\item \code{grk_reindent_tabs_text}: Re-indent text using tabs

\item \code{grk_reindent_tabs_file}: Re-indent a file using tabs

\item \code{grk_reindent_tabs_dir}: Re-indent a directory using tabs

\item \code{grk_reindent_tabs_pkg}: Re-indent a package using tabs

\item \code{grk_reindent_spaces_text}: Re-indent text using spaces

\item \code{grk_reindent_spaces_file}: Re-indent a file using spaces

\item \code{grk_reindent_spaces_dir}: Re-indent a directory using spaces

\item \code{grk_reindent_spaces_pkg}: Re-indent a package using spaces
}}

