% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{grk_style}
\alias{grk_style}
\alias{use_grk_style}
\alias{grk_style_transformer}
\alias{grk_style_text}
\alias{grk_style_file}
\alias{grk_style_dir}
\alias{grk_style_pkg}
\title{The grkstyle Code Style}
\usage{
use_grk_style()

grk_style_transformer(..., use_tabs = getOption("grkstyle.use_tabs", NULL))

grk_style_text(text, ...)

grk_style_file(path, ...)

grk_style_dir(path, ...)

grk_style_pkg(pkg = ".", ...)
}
\arguments{
\item{...}{Arguments passed to underling \pkg{styler} functions (identified
by removing the \code{grk_} prefix), except for \code{transformers}, which is set to
the \code{grk_style_transformer()} internally.}

\item{use_tabs}{Should a single tab be used for indentation? \pkg{grkstyle}
functions will by default follow the \code{grkstyle.use_tabs} option, if set, or
will attempt to look up the project settings in the current RStudio
project. You can choose to use tabs or spaces for a project by changing the
\emph{Insert spaces for tab} option in the Code Editing panel of the Project
Options settings.

If not otherwise set via the R option or the RStudio project option, tabs
are used. Tabs are recommended because they are
\href{https://alexandersandberg.com/articles/default-to-tabs-instead-of-spaces-for-an-accessible-first-environment/}{more accessible}.

You can opt into indentation by two spaces by setting the
\code{grkstyle.use_tabs} option:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# two spaces
options(grkstyle.use_tabs = FALSE)

# equivalently, but more precise
options(grkstyle.use_tabs = list(indent_by = 2, indent_character = " "))
}\if{html}{\out{</div>}}}

\item{text}{A character vector with text to style.}

\item{path}{A character vector with paths to files to style.}

\item{pkg}{Path to a (subdirectory of an) R package.}
}
\description{
Use \code{\link[styler:style_text]{styler::style_text()}} to format code according to the unofficial
\pkg{grkstyle} style guide. Follows the
\href{https://style.tidyverse.org}{tidyverse style guide} as implemented by
\code{\link[styler:tidyverse_style]{styler::tidyverse_style()}}, but with additional style rules that enforce
consistent line breaks inside function calls and left-justification of
function arguments in function definitions.
}
\section{Functions}{
\itemize{
\item \code{use_grk_style}: Set the \pkg{grkstyle} style as the default style for
\pkg{styler} addins.

\item \code{grk_style_transformer}: A code transformer for use with \code{\link[styler:style_text]{styler::style_text()}}

\item \code{grk_style_text}: Style text using the \pkg{grkstyle} code style

\item \code{grk_style_file}: Style a file using the \pkg{grkstyle} code style

\item \code{grk_style_dir}: Style a directory using the \pkg{grkstyle} code style

\item \code{grk_style_pkg}: Style a package using the \pkg{grkstyle} code style
}}

\section{Using the \pkg{grkstyle} code style}{
 You can set the \pkg{grkstyle}
code style as the default code style for \pkg{styler} (and its associated
RStudio addins, like "Style active file" and "Style selection") by calling
\code{grkstyle::use_grk_style()}. If you would rather set this option globally
for all session, you can add the following to your \code{.Rprofile}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(styler.addins_style_transformer = "grkstyle::grk_style_transformer()")
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
use_grk_style()
# Use styler addins
styler:::style_selection()
}

ex_code <- "mtcars \%>\% mutate(mpg = mpg * 2,\n\t\t     cyl = paste(cyl)) \%>\% head()"
cat(ex_code)

grk_style_text(ex_code)
}
